+++ lib/src/printer/empty.dart
@@ -1,3 +1,4 @@
+import 'object/object.dart';
 import 'printer.dart';
 
 extension EmptyPrinterExtension<T> on Printer<Iterable<T>> {
@@ -17,4 +18,9 @@ class EmptyPrinter<T> extends Printer<Iterable<T>> {
   @override
   void printOn(Iterable<T> object, StringBuffer buffer) =>
       object.isEmpty ? buffer.write(label) : printer.printOn(object, buffer);
+
+  @override
+  ObjectPrinter get toStringPrinter => super.toStringPrinter
+    ..addValue(printer, name: 'printer')
+    ..addValue(label, name: 'label');
 }