+++ lib/src/printer/number/fixed.dart
@@ -1,5 +1,6 @@
 import '../../../math.dart';
 import '../builder.dart';
+import '../object/object.dart';
 import '../printer.dart';
 import '../string/pad.dart';
 import '../string/separate.dart';
@@ -107,4 +108,17 @@ class FixedNumberPrinter<T extends num> extends Printer<T> {
     final result = formatDigits(digits, characters);
     _fraction.printOn(result, buffer);
   }
+
+  @override
+  ObjectPrinter get toStringPrinter => super.toStringPrinter
+    ..addValue(accuracy, name: 'accuracy')
+    ..addValue(base, name: 'base')
+    ..addValue(characters, name: 'characters')
+    ..addValue(delimiter, name: 'delimiter')
+    ..addValue(infinity, name: 'infinity')
+    ..addValue(nan, name: 'nan')
+    ..addValue(padding, name: 'padding')
+    ..addValue(precision, name: 'precision')
+    ..addValue(separator, name: 'separator')
+    ..addValue(sign, name: 'sign');
 }