+++ lib/src/printer/number/human.dart
@@ -1,4 +1,5 @@
 import '../../../math.dart';
+import '../object/object.dart';
 import '../printer.dart';
 import 'fixed.dart';
 import 'sign.dart';
@@ -248,4 +249,21 @@ class HumanNumberPrinter<T extends num> extends Printer<T> {
       return (value.abs().log() / unitBase.log()).floor();
     }
   }
+
+  @override
+  ObjectPrinter get toStringPrinter => super.toStringPrinter
+    ..addValue(base, name: 'base')
+    ..addValue(characters, name: 'characters')
+    ..addValue(delimiter, name: 'delimiter')
+    ..addValue(infinity, name: 'infinity')
+    ..addValue(nan, name: 'nan')
+    ..addValue(padding, name: 'padding')
+    ..addValue(precision, name: 'precision')
+    ..addValue(separator, name: 'separator')
+    ..addValue(sign, name: 'sign')
+    ..addValue(unitBase, name: 'unitBase')
+    ..addValue(unitOffset, name: 'unitOffset')
+    ..addValue(unitPrefix, name: 'unitPrefix')
+    ..addValue(unitSeparator, name: 'unitSeparator')
+    ..addValue(units, name: 'units');
 }