+++ lib/src/printer/object/field.dart
@@ -1,4 +1,5 @@
 import '../printer.dart';
+import 'object.dart';
 
 /// Abstract field description.
 abstract class FieldPrinter<T> extends Printer<T> {
@@ -7,4 +8,8 @@ abstract class FieldPrinter<T> extends Printer<T> {
 
   /// Returns true if the field should be omitted from display.
   bool isOmitted(T object);
+
+  @override
+  ObjectPrinter get toStringPrinter =>
+      super.toStringPrinter..addValue(name, name: 'name');
 }