+++ lib/src/printer/pluggable.dart
@@ -1,4 +1,5 @@
 import '../functional/types/mapping.dart';
+import 'object/object.dart';
 import 'printer.dart';
 
 /// Evaluates the callback with the value to retrieve the string.
@@ -9,4 +10,8 @@ class PluggablePrinter<T> extends Printer<T> {
 
   @override
   void printOn(T object, StringBuffer buffer) => buffer.write(callback(object));
+
+  @override
+  ObjectPrinter get toStringPrinter =>
+      super.toStringPrinter..addValue(callback, name: 'callback');
 }