+++ lib/src/printer/string/truncate.dart
@@ -1,5 +1,6 @@
 import 'package:characters/characters.dart';
 
+import '../object/object.dart';
 import '../printer.dart';
 
 extension TruncatePrinterExtension<T> on Printer<T> {
@@ -12,13 +13,25 @@ extension TruncatePrinterExtension<T> on Printer<T> {
       TruncateRightPrinter<T>(this, width, ellipsis);
 }
 
-/// Truncates the string from the left side if it is longer than width.
-class TruncateLeftPrinter<T> extends Printer<T> {
-  const TruncateLeftPrinter(this.printer, this.width, this.ellipsis);
+/// Truncates the string if it is longer than width.
+abstract class TruncatePrinter<T> extends Printer<T> {
+  const TruncatePrinter(this.printer, this.width, this.ellipsis);
 
   final Printer<T> printer;
   final int width;
   final String ellipsis;
+
+  @override
+  ObjectPrinter get toStringPrinter => super.toStringPrinter
+    ..addValue(printer, name: 'printer')
+    ..addValue(width, name: 'width')
+    ..addValue(ellipsis, name: 'ellipsis');
+}
+
+/// Truncates the string from the left side if it is longer than width.
+class TruncateLeftPrinter<T> extends TruncatePrinter<T> {
+  const TruncateLeftPrinter(Printer<T> printer, int width, String ellipsis)
+      : super(printer, width, ellipsis);
 
   @override
   void printOn(T object, StringBuffer buffer) {
@@ -33,12 +46,9 @@ class TruncateLeftPrinter<T> extends Printer<T> {
 }
 
 /// Truncates the string from the right side if it is longer than width.
-class TruncateRightPrinter<T> extends Printer<T> {
-  const TruncateRightPrinter(this.printer, this.width, this.ellipsis);
-
-  final Printer<T> printer;
-  final int width;
-  final String ellipsis;
+class TruncateRightPrinter<T> extends TruncatePrinter<T> {
+  const TruncateRightPrinter(Printer<T> printer, int width, String ellipsis)
+      : super(printer, width, ellipsis);
 
   @override
   void printOn(T object, StringBuffer buffer) {