+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 3.4.0 (unpublished)
+## 3.4.0
 * Remove deprecated `hashAll` and `hash1`, `hash2`, `hash3`, ... methods.
 * Add `ObjectPrinter.addValue` for more dynamic object printing.
 * Add `ToStringPrinter` mixin for simpler use of `ObjectPrinter`.